Split gragh and viewer
