+++ app/[lng]/debugger/debugger.tsx
@@ -1,434 +1,18 @@
 'use client';
 
-import { useEffect, useRef, useState, useCallback, ChangeEventHandler, RefObject, SubmitEventHandler } from "react";
-import Chart from "react-apexcharts";
-import { advFilterLog, clearLineStr, LogParser } from "./log_parser";
+import { useRef, useState, useCallback, ChangeEventHandler } from "react";
+import { CHUNK_SIZE, LINES_PER_TICK, LogParser, readBlobAsText, yieldToMain, ZoomType } from "./log_parser";
 import './debugger.css';
-import type { ApexOptions } from 'apexcharts';
 import { useT } from "@/app/i18n/client";
-import { MagnifyingGlassCircleIcon } from "@heroicons/react/24/outline";
-
-interface ZoomRange {
-    min: number;
-    max: number;
-}
-
-type ZoomType = {
-    mindate?: Date,
-    maxdate?: Date,
-    enabled: boolean
-}
+import dynamic from 'next/dynamic';
+const LogViewer = dynamic(() => import("./LogViewer"));
+const LogGraph = dynamic(() => import("./LogGraph"));
 
 type LogData = {
     size: number,
     read_size: number,
     climates: string[],
     state: "empty" | "running" | "finished",
-}
-
-
-const CHUNK_SIZE = 2 * 1024 * 1024; // 2 Mo par chunk
-const LINES_PER_TICK = 500;         // lignes traitées avant de céder le thread + mise à jour de la progression
-
-type ZoomCallback = (datemin: Date, datemax: Date) => void
-
-
-/** Lit un Blob en texte de façon asynchrone */
-function readBlobAsText(blob: Blob): Promise<string> {
-    return new Promise((resolve, reject) => {
-        const reader = new FileReader();
-        reader.onload = () => resolve(reader.result as string);
-        reader.onerror = () => reject(reader.error);
-        reader.readAsText(blob);
-    });
-}
-
-
-/** Cède le contrôle au navigateur (évite le freeze) */
-function yieldToMain(): Promise<void> {
-    return new Promise((resolve) => setTimeout(resolve, 0));
-}
-
-function dateSort(a: { x: Date }, b: { x: Date }): number {
-    return a.x.getTime() - b.x.getTime()
-}
-
-const Graph: React.FC<{ logfile: RefObject<LogParser>, selectedThermostat: string, onZoomChange?: ZoomCallback, zoom?: ZoomType, onZoomReset?: () => void }> = ({ logfile, selectedThermostat, onZoomChange, zoom, onZoomReset }) => {
-    const { t, i18n } = useT('analyzer');
-    const [isDark, setDark] = useState<boolean>(() => {
-        if (typeof window === 'undefined') return false
-        const mq = window.matchMedia('(prefers-color-scheme: dark)')
-        return document.documentElement.classList.contains('dark') || mq.matches
-    });
-    const series: ApexOptions['series'] = [
-        {
-            name: t('graph.target'),
-            data: logfile.current.getThermostat(selectedThermostat)?.target_temps.map(event => ({
-                x: event.timestamp,
-                y: event.value,
-            }))
-                .sort((a, b) => a.x.getTime() - b.x.getTime()) || [],
-            type: 'area',
-        },
-        {
-            name: t('graph.room_temp'),
-            data: logfile.current.getThermostat(selectedThermostat)?.room_temps.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || []
-        },
-        {
-            name: t('graph.ext_temp'),
-            data: logfile.current.getThermostat(selectedThermostat)?.ext_temps.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || []
-        },
-        {
-            name: t('graph.underlying_setpoint'),
-            data: logfile.current.getThermostat(selectedThermostat)?.underlying_setpoints.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || []
-        },
-        {
-            name: t('graph.underlying_temp'),
-            data: logfile.current.getThermostat(selectedThermostat)?.underlying_temps.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || []
-        },
-        {
-            name: t('graph.regulated_temp'),
-            data: logfile.current.getThermostat(selectedThermostat)?.regulated_temps.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || []
-        },
-    ];
-
-    const series_features: ApexOptions['series'] = [
-        {
-            name: t('graph.window'),
-            data: logfile.current.getThermostat(selectedThermostat)?.window_state.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || []
-        },
-        {
-            name: t('graph.safety'),
-            data: logfile.current.getThermostat(selectedThermostat)?.safety_state.map(event => ({ x: event.timestamp, y: event.value })).sort(dateSort) || [],
-            color: 'red'
-        },
-        {
-            name: 'Central Boiler',
-            data: logfile.current.central_boiler.map(event => ({ x: event.timestamp, y: event.value })).sort((a, b) => a.x.getTime() - b.x.getTime())
-        }
-    ];
-
-    // Find last timestamp of all series then add this timestamp to other series
-    const end_timestamps = series.map((x, i) => {
-        if (x.data.length == 0)
-            return { serie: i, last: new Date(0) }
-        const lastElement = x.data[x.data.length - 1];
-        const lastValue = typeof lastElement === 'object' && lastElement !== null && 'x' in lastElement ? lastElement.x : new Date(0);
-        return { serie: i, last: lastValue }
-    })
-        .sort((a, b) => a.last.getTime() - b.last.getTime())
-
-    const last_timestamp = end_timestamps[end_timestamps.length - 1]
-
-    for (const [i, serie] of series.entries()) {
-        if (i == last_timestamp.serie || serie.data.length == 0)
-            continue;
-        const lastElement = serie.data[serie.data.length - 1];
-        if (typeof lastElement === 'object' && lastElement !== null && 'y' in lastElement) {
-            (serie.data as { x: Date, y: number }[]).push({ x: last_timestamp.last, y: lastElement.y });
-        }
-    }
-
-    for (const serie of series_features) {
-        if (serie.data.length == 0)
-            continue;
-        const lastElement = serie.data[serie.data.length - 1];
-        if (typeof lastElement === 'object' && lastElement !== null && 'y' in lastElement) {
-            (serie.data as { x: Date, y: number }[]).push({ x: last_timestamp.last, y: lastElement.y });
-        }
-    }
-
-    // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-    function handleZoomChange(chartContext: any, { xaxis }: { xaxis: ZoomRange }) {
-        const dateMin = new Date(xaxis.min);
-        const dateMax = new Date(xaxis.max);
-        if (onZoomChange != undefined) onZoomChange(dateMin, dateMax);
-    }
-
-    console.log('Series', series);
-    console.log('Other states', logfile.current.getThermostat(selectedThermostat)?.window_state);
-
-    const xaxis: ApexOptions['xaxis'] = {
-        type: 'datetime',
-        title: { text: 'Time' },
-        labels: { format: 'HH:mm:ss', datetimeUTC: false },
-        min: zoom?.enabled ? zoom?.mindate?.getTime() : undefined,
-        max: zoom?.enabled ? zoom?.maxdate?.getTime() : undefined,
-        tooltip: {
-            formatter: (value) => new Intl.DateTimeFormat(i18n.language, { dateStyle: "medium", timeStyle: "short" }).format(new Date(value)),
-        }
-    };
-
-    useEffect(() => {
-        const mq = window.matchMedia('(prefers-color-scheme: dark)')
-        const handler = (e: MediaQueryListEvent) => setDark(e.matches)
-        mq.addEventListener('change', handler)
-        return () => mq.removeEventListener('change', handler)
-    }, []);
-
-    const colors = isDark ?
-        [
-            '#FFC107', // Jaune/Orange
-            '#4DBAFA', // Bleu clair
-            '#26E7A6', // Menthe
-            '#C49384', // Marron clair
-            '#9B86E0', // Violet
-            '#7986CB', // Indigo
-            '#5EC5ED', // Cyan
-            '#90A4AE', // Gris-bleu
-            '#EC6B87', // Rose/Rouge
-            '#C9BCB1', // Beige/Sable
-            '#76D7D0'  // Turquoise
-        ]
-        : ['#FEB019', '#008FFB', '#00E396', '#8D5B4C', '#775DD0', '#3f51b5', '#33b2df', '#546E7A', 'd4526e', '#A5978B', '#4ecdc4'];
-
-    return (
-        <div className="w-full">
-            <Chart
-                className="chart"
-                options={{
-                    chart: {
-                        id: 'log-chart',
-                        group: 'log-charts',
-                        type: 'line',
-                        zoom: {
-                            enabled: true,
-                            allowMouseWheelZoom: false,
-                        },
-                        events: {
-                            zoomed: handleZoomChange,
-                            beforeResetZoom: onZoomReset,
-                        },
-                    },
-                    fill: {
-                        opacity: [0.3, 1, 1, 1, 1, 1]
-                    },
-                    theme: {
-                        mode: isDark ? 'dark' : 'light',
-                    },
-                    tooltip: {
-                        intersect: false,
-                        shared: true,
-                        x: {
-                            show: false
-                        },
-                        custom: ({ seriesIndex, dataPointIndex, w }) => {
-                            // Timestamp du point survolé
-                            const hoveredTs = w.globals.seriesX[seriesIndex][dataPointIndex];
-
-                            // Pour chaque série, on cherche la dernière valeur <= hoveredTs
-                            const rows = w.globals.seriesNames
-                                .map((name: string, i: number) => {
-                                    const xArr = w.globals.seriesX[i];
-                                    const yArr = w.globals.series[i];
-
-                                    // Dernier index dont le timestamp est <= hoveredTs
-                                    let lastVal: null | number = null;
-                                    for (let j = 0; j < xArr.length; j++) {
-                                        if (xArr[j] <= hoveredTs) {
-                                            lastVal = yArr[j];
-                                        }
-                                    }
-
-                                    const color = w.globals.colors[i];
-                                    const display = lastVal !== null ? lastVal.toFixed(2) : "–";
-
-                                    return `
-                                  <div style="display:flex; align-items:center; gap:6px; padding:2px 0">
-                                    <span style="width:10px;height:10px;border-radius:50%;
-                                                 background:${color};display:inline-block"></span>
-                                    <span>${name} :</span>
-                                    <strong>${display}</strong>
-                                  </div>`;
-                                })
-                                .join("");
-
-                            return `
-                              <div style="padding:10px 14px; font-size:13px; font-family:sans-serif">
-                                ${rows}
-                              </div>`;
-                        },
-                    },
-                    xaxis,
-                    yaxis: {
-                        labels: {
-                            formatter: (val: number) => val.toFixed(2)
-                        }
-                    },
-                    stroke: { curve: 'stepline', width: [2, 2, 1, 1, 1.5, 1.5] },
-                    title: { text: t('graph.title_logs'), align: 'left' },
-                    colors,
-                }}
-                series={series}
-                type="line"
-                height={350}
-            />
-            <Chart
-                type="area"
-                className="chart"
-                height={200}
-                options={{
-                    chart: {
-                        id: 'feat-chart',
-                        // group: 'log-charts',
-                        type: 'line',
-                        zoom: {
-                            enabled: true,
-                            allowMouseWheelZoom: false,
-                        },
-                        events: {
-                            zoomed: handleZoomChange,
-                            beforeResetZoom: onZoomReset
-                        },
-                    },
-                    title: {
-                        text: t('graph.title_feature')
-                    },
-                    legend: {
-                        show: true,
-                    },
-                    dataLabels: {
-                        enabled: false,
-                    },
-                    theme: {
-                        mode: isDark ? 'dark' : 'light',
-                    },
-                    xaxis,
-                    stroke: { curve: 'stepline', width: 2 },
-                    yaxis: {
-                        labels: {
-                            formatter: (val) => {
-                                if (val == 1) return "ON";
-                                if (val == 0) return "OFF";
-                                if (val == -1) return "BYPASS";
-                                return val.toString();
-                            }
-                        },
-                        stepSize: 1,
-                        min: -1,
-                        max: 1,
-                    }
-                }}
-                series={series_features}
-            />
-        </div>
-    );
-}
-
-const EditorV2: React.FC<{
-    parser: RefObject<LogParser>,
-    climate?: string,
-    file_input: RefObject<HTMLInputElement | null>,
-    zoom: ZoomType
-}> = ({ parser, climate, file_input, zoom }) => {
-    const inref = useRef<HTMLDivElement>(null);
-    const [adv_filter, setAdvFilter] = useState<{
-        vtherm_only: boolean,
-        txt?: string
-    }>({
-        vtherm_only: false,
-        txt: undefined
-    })
-
-    const write_log = useCallback(async (file: File, edit_div: HTMLDivElement) => {
-        let offset = 0;
-        let leftover = "";
-
-        edit_div.innerText = ''
-        const filter_txt = typeof adv_filter.txt === "string" && adv_filter.txt.length > 0 ? RegExp(adv_filter.txt) : null
-
-        while (offset < file.size) {
-            const chunk = file.slice(offset, offset + CHUNK_SIZE);
-            const text = await readBlobAsText(chunk);
-            offset += chunk.size;
-
-            // Découper en lignes en conservant le reste incomplet
-            const combined = leftover + text;
-            const lines = combined.split("\n");
-            leftover = lines.pop() ?? ""; // la dernière portion peut être incomplète
-
-
-            // Traiter les lignes par petits lots pour ne pas bloquer le thread
-            for (let i = 0; i < lines.length; i++) {
-
-                //parseLine(lines[i]);
-                const line = clearLineStr(lines[i])
-                const skip = line.match(/^[^[]*\[[^]*\]\W*$/)
-                if (skip) {
-                    continue;
-                }
-
-                if (adv_filter.vtherm_only && advFilterLog(line)) {
-                    continue;
-                }
-
-                if (filter_txt) {
-                    if (!filter_txt.test(line)) {
-                        continue;
-                    }
-                }
-
-                const { climate: log_climate, level, date, txt } = parser.current.getLogTextInfos(line)
-
-                if (log_climate == '' || climate == log_climate) {
-                    if (!zoom.enabled || !date || (zoom.mindate && date > zoom.mindate && zoom.maxdate && date < zoom.maxdate)) {
-                        const pl = document.createElement('p');
-                        pl.appendChild(document.createTextNode(txt));
-                        if (level == 'INFO') pl.className = 'info';
-                        if (level == 'WARNING') pl.className = 'warning';
-                        if (level == 'ERROR') pl.className = 'error';
-                        if (!date) pl.className = 'nodata';
-                        edit_div.appendChild(pl)
-                    }
-                }
-
-                // Cède le contrôle + met à jour la progression tous les LINES_PER_TICK
-                if ((i + 1) % LINES_PER_TICK === 0) {
-                    // TOTO: updateProgress();
-                    await yieldToMain();
-                }
-            }
-        }
-    }, [zoom, climate, adv_filter, parser]);
-
-    useEffect(() => {
-        if (!inref.current) {
-            console.warn('EDITOR - Div not ready');
-            return;
-        }
-        if (!file_input?.current?.files?.length) {
-            console.warn('EDITOR - File not ready');
-            return;
-        }
-        if (!climate) {
-            console.warn('EDITOR - No climate selected');
-            return;
-        };
-        write_log(file_input.current.files[0], inref.current);
-    }, [climate, zoom, file_input, adv_filter, write_log])
-
-    const onSearch: SubmitEventHandler<HTMLFormElement> = (evt) => {
-        evt.preventDefault()
-        setAdvFilter(e => ({ ...e, txt: evt.target['txt'].value }))
-    }
-
-    return <div className="w-full">
-        <div className="py-2 border rounded-2xl my-4 px-3">
-            <h2 className="text-xl border-b">Filters</h2>
-            <form onSubmit={onSearch} className="pt-4 flex flex-col">
-                <div className="flex">
-                    <input type="checkbox" id="log_filter" checked={adv_filter.vtherm_only} onChange={() => setAdvFilter(e => ({ ...e, vtherm_only: !adv_filter.vtherm_only }))} />
-                    <label htmlFor="log_filter" className="px-2 block" >Enable advance log filtering</label>
-                </div>
-                <div>
-                    <div className="pr-3 inline">Search in logs:</div>
-                    <input type="text" name="txt" placeholder="Regex search" className="w-2xs max-w-full bg-slate-100 px-2 py-1 rounded" />
-                </div>
-                <button type="submit" className="inline w-fit bg-green-300 dark:bg-green-400 dark:text-black rounded-full px-6 my-3 py-2"><MagnifyingGlassCircleIcon className="h-4 cursor-pointer inline" /> Search</button>
-            </form>
-        </div>
-        <div ref={inref} className="debugger"></div>
-    </div>
 }
 
 const Debugger: React.FC = () => {
@@ -577,8 +161,8 @@ const Debugger: React.FC = () => {
                 {logData.size > 0 && <button type="reset" className="bg-red-100 dark:bg-red-400 dark:text-black hover:bg-red-200 transition-all duration-400 cursor-pointer font-bold rounded-full px-4 text-red-800 border-red-500 border-solid border">{t('reset')}</button>}
                 {logData.size > 0 && <button type="submit" className="bg-blue-100 dark:bg-blue-400 dark:text-black hover:bg-blue-200 transition-all duration-400 cursor-pointer font-bold rounded-full px-4 text-blue-800 border-blue-500 border-solid border">{t('reload')}</button>}
             </form>
-            {logData.state == 'finished' && selectedThermostat && <Graph logfile={parser} selectedThermostat={selectedThermostat} zoom={zoom} onZoomChange={onZoomChange} onZoomReset={onZoomReset} />}
-            {logData.state == 'finished' && selectedThermostat && <EditorV2 climate={selectedThermostat} parser={parser} file_input={fileInputRef} zoom={zoom} />}
+            {logData.state == 'finished' && selectedThermostat && <LogGraph logfile={parser} selectedThermostat={selectedThermostat} zoom={zoom} onZoomChange={onZoomChange} onZoomReset={onZoomReset} />}
+            {logData.state == 'finished' && selectedThermostat && <LogViewer climate={selectedThermostat} parser={parser} file_input={fileInputRef} zoom={zoom} />}
         </div>
     );
 }
